Add files via upload
The management of the attendance can be a great burden on the teachers if it is done by hand. To resolve this problem, smart and auto attendance management system is being utilized. By utilizing this framework, the problem of proxies and students being marked present even though they are not physically present can easily be solved. This system marks the attendance using live video stream. The frames are extracted from video using OpenCV. The main implementation steps used in this type of system are face detection and recognizing the detected face, for which face-recognition library using dlib is used. After these, the connection of recognized faces ought to be conceivable by comparing with the database containing students’ faces. This model will be a successful technique to manage the attendance of students.